Add include and library paths so the test example compiles
Don't know if this is new in conan2, something changed in cmake, or
it's down to changing the way the main repo was installed, but the
include and linking paths were not correct and caused compilation of the
test program to fail. This is now fixed.

We have also updated the minium version of cmake and reordered where it
is defined to remove warnings at configuration time.